Moved more transformations to hwutils
